+++ published/kinetic_request_ce_cascading_attribute_value_retrieve_v2/handler/init.rb
@@ -24,14 +24,21 @@ class KineticRequestCeCascadingAttributeValueRetrieveV2
   end
 
   def execute
+    space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     begin
-      space_slug      = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
       error_handling  = @parameters["error_handling"]
       error_message   = nil
 
       api_username    = URI.encode(@info_values["api_username"])
       api_password    = @info_values["api_password"]
-      api_server      = @info_values["api_server"]
       start_context   = @parameters["start_context"]
       end_context     = @parameters["end_context"]
       prop_to_find    = @parameters["prop_to_find"]
@@ -39,7 +46,7 @@ class KineticRequestCeCascadingAttributeValueRetrieveV2
       kapp_slug       = @parameters["kapp_slug"].empty? ? nil : @parameters["kapp_slug"]
       form_slug       = @parameters["form_slug"].empty? ? nil : @parameters["form_slug"]
       submission_id   = @parameters["submission_id"].empty? ? nil : @parameters["submission_id"]
-      api_route       = "#{api_server}/#{space_slug}/app/api/v1/"
+      api_route       = "#{server}/app/api/v1/"
       valid_route     = false
 
       ######################################################