+++ published/kinetic_request_ce_datastore_submission_activity_update_v1/handler/init.rb
@@ -25,15 +25,22 @@ class KineticRequestCeDatastoreSubmissionActivityUpdateV1
 
   def execute
     space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     error_handling  = @parameters["error_handling"]
     error_message = nil
 
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
 
     begin
-      api_route = "#{api_server}/#{space_slug}/app/api/v1/datastore/submissions/#{@parameters["submission_id"]}/activities/#{@parameters["submission_activity_id"]}"
+      api_route = "#{server}/app/api/v1/datastore/submissions/#{@parameters["submission_id"]}/activities/#{@parameters["submission_activity_id"]}"
 
       resource = RestClient::Resource.new(api_route, { :user => api_username, :password => api_password })