+++ published/kinetic_request_ce_form_retrieve_v1/handler/init.rb
@@ -24,15 +24,22 @@ class KineticRequestCeFormRetrieveV1
   end
 
   def execute
+    space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
     kapp_slug       = @parameters["kapp_slug"]
     form_slug       = @parameters["form_slug"]
-    space_slug      = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
     error_handling  = @parameters["error_handling"]
 
-    api_route = "#{api_server}/#{space_slug}/app/api/v1/kapps/#{kapp_slug}/forms/#{form_slug}?include=details,attributes,fields"
+    api_route = "#{server}/app/api/v1/kapps/#{kapp_slug}/forms/#{form_slug}?include=details,attributes,fields"
 
     puts "API ROUTE: #{api_route}" if @enable_debug_logging