+++ published/kinetic_request_ce_space_retrieve_v1/handler/init.rb
@@ -24,14 +24,21 @@ class KineticRequestCeSpaceRetrieveV1
   end
 
   def execute
+    space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
     kapp_slug       = @parameters["kapp_slug"]
-    space_slug      = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
     error_handling  = @parameters["error_handling"]
 
-    api_route = "#{api_server}/#{space_slug}/app/api/v1/space?include=details,attributes"
+    api_route = "#{server}/app/api/v1/space?include=details,attributes"
 
     resource = RestClient::Resource.new(api_route, { :user => api_username, :password => api_password })