+++ unpublished/kinops_legacy/kinetic_request_ce_submission_retrieve_by_query_v1/test/simple_input.rb
@@ -1,0 +1,15 @@
+{
+  'info' => {
+    'api_server' => 'https://localhost:8080/kinetic',
+    'api_username' => 'user@kineticdata.com',
+    'api_password' => 'password',
+    'space_slug' => 'acme'
+  },
+  'parameters' => {
+    'error_handling' => 'Error Message',
+    'space_slug' => '',
+    'kapp_slug' => 'admin',
+    'form_slug' => 'message-recipient-and-content',
+    'query' => 'values[Record%20Type]=Content&values[Notification%20Name]=Sample'
+  }
+}