+++ unpublished/kinops_legacy/kinetic_request_ce_user_retrieve_from_list_v1/test/simple_input.rb
@@ -1,0 +1,12 @@
+{
+  'info' => {
+    'api_server' => 'https://localhost:8080/kinetic',
+    'api_username' => 'user@kineticdata.com',
+    'api_password' => 'password',
+    'space_slug' => 'acme'
+  },
+  'parameters' => {
+    'space_slug' => 'matt-playground',
+    'usernames' => 'matthew.howe@kineticdata.com,matthew.howe.2x@kineticdata.com'
+  }
+}