+++ published/kinetic_request_ce_cascading_attribute_value_retrieve_v2/CHANGELOG.md
@@ -1,3 +1,7 @@
+Kinetic Request CE Cascading Attribute Value Retrieve V2.1 (2018-05-25)
+* API Server Info Value changed to allow ${space} in the url for subdomain support
+(ie. https://${space}.localhost:8080/kinetic)
+
 Kinetic Request CE Cascading Attribute Value Retrieve V2 (2017-08-18)
 * Removed menus from start and end contexts to allow values to be passed in as variables (from
 routine input)